+++ package.json
@@ -2,20 +2,21 @@
   "name": "ws_task",
   "version": "1.0.0",
   "description": "Template for the WebSocket remote control task",
-  "main": "index.js",
-  "type": "module",
+  "main": "index.ts",
   "scripts": {
-    "start": "node ./index.js",
-    "start:dev": "nodemon ./index.js",
+    "start": "node ./index.ts",
+    "start:dev": "nodemon ./index.ts",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "Andrei Auchynnikau",
   "keywords": [],
   "license": "ISC",
   "devDependencies": {
+    "@types/ws": "^8.5.10",
     "nodemon": "^2.0.16"
   },
   "dependencies": {
+    "ts-node": "^10.9.2",
     "ws": "^8.16.0"
   }
 }
