+++ src/websocket_server/index.ts
@@ -1,0 +1,9 @@
+import { WebSocketServer } from "ws";
+
+export function runWebSocketServer(websocketPort: number) {
+  const server = new WebSocketServer({ port: websocketPort });
+
+  server.on("connection", () => {
+    console.log("New client connected!");
+  });
+}
